backend/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-CORS==6.0.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
Werkzeug==3.1.3